Update Optimus Zarr Output with EmptyDrops(GL-576) (#248)
Update Optimus ZARR output to include emptydrops data. In the process, added versioning to the ZARR file to ease parsing and flatened the structure. A new image for this ZARR conversion was added as well as tests. The relevant documentation under docs was updated.
