+++ docker/zarr-output/requirements.txt
@@ -1,0 +1,36 @@
+asciitree==0.3.3
+atomicwrites==1.3.0
+attrs==19.1.0
+blosc==1.8.1
+certifi==2019.6.16
+chardet==3.0.4
+Click==7.0
+Crimson==0.4.0
+cycler==0.10.0
+fasteners==0.15
+HTSeq==0.11.2
+idna==2.6
+importlib-metadata==0.19
+kiwisolver==1.1.0
+matplotlib==3.1.1
+monotonic==1.5
+more-itertools==7.2.0
+numcodecs==0.6.3
+packaging==19.1
+pandas==0.25.0
+pluggy==0.12.0
+py==1.8.0
+pyparsing==2.4.2
+pysam==0.15.3
+pytest==5.1.1
+python-dateutil==2.8.0
+pytz==2019.2
+PyYAML==5.1.2
+requests==2.18.4
+scipy==1.3.1
+sctools==0.1.6
+six==1.12.0
+urllib3==1.22
+wcwidth==0.1.7
+zarr==2.2.0
+zipp==0.5.2
