+++ docker/zarr-output/test/README.md
@@ -1,0 +1,1 @@
+The `run_test.sh` script will compile the docker image as a test image and then run a test command on it verifying that the command works. The output of this command can then be found in the `testout` directory.
